nl2controller/requirements.txt
# NL2Controller Dependencies
# ===========================

# Core Dependencies
python-dotenv>=1.0.0
pydantic>=2.0.0
numpy>=1.20.0
requests>=2.25.0

# LLM Integration (for lifting stage)
openai>=1.0.0

# Formal synthesis with TuLiP
tulip>=1.4.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0
